refactor: move invalidate function to cache module
